Never remove topmost backend directory
Never remove topmost backend directory.

+++ src/s3ql/fsck.py
@@ -929,7 +929,7 @@ class Fsck(object):
 
         log.info('Checking for temporary objects and empty directories (backend)...')
 
-        empty_dirs=0
+        empty_dirs = 0
         for (path, dirnames, filenames) in os.walk(plain_backend.prefix, topdown=False):
             for name in filenames:
                 if not re.search(r'^[^#]+#[0-9]+--?[0-9]+\.tmp$', name):
@@ -938,11 +938,14 @@ class Fsck(object):
                 self.found_errors = True
                 self.log_error("removing temporary file %s", name)
                 os.unlink(os.path.join(path, name))
-            try:
-                empty_dirs+=1
-                os.rmdir(path)
-            except OSError:
-                empty_dirs-=1
+                
+            if path != plain_backend.prefix:
+                try:
+                    empty_dirs += 1
+                    os.rmdir(path)
+                except OSError:
+                    empty_dirs -= 1
+                    
         log.info('Removed %d empty directories', empty_dirs)
 
     def check_objects_id(self):
